Freeze/Unfreeze completed

+++ CommonTypes/BaseProcess.cs
@@ -19,12 +19,16 @@ namespace CommonTypes
         // system status (frozen, unfrozen)
         public Status Status { get; protected set; }
 
+        //List of actions saved when process state is frozen
+        public List<String[]> FrozenStateList { get; set; }
+
         protected BaseProcess(string processName, string processUrl, string puppetMasterUrl)
         {
             ProcessName = processName;
             Url = processUrl;
             // connects to this site's puppetMaster
             PuppetMaster = (IProcessMaster)Activator.GetObject(typeof(IProcessMaster), puppetMasterUrl);
+            FrozenStateList = new List<string[]>();
         }
 
         /// <summary>
@@ -54,28 +58,81 @@ namespace CommonTypes
             return brokerUrls;
         }
 
+        public virtual void ProcessFrozenListCommands()
+        {
+            foreach (String[] command in FrozenStateList)
+            { 
+                switch (command[0])
+                {
+                    case "Status":
+                        Console.Out.WriteLine("Printing status information:\r\nSubscriptions, etc...");
+                        break;
+
+                    case "Crash":
+                        Console.WriteLine("Crashing");
+                        Process.GetCurrentProcess().Kill();
+                        break;
+
+                }
+            }
+        }
+
         public virtual void DeliverCommand(string[] command)
         {
-            switch (command[0])
+            if (Status == Status.Frozen)
             {
-                case "Status":
-                    Console.Out.WriteLine("Printing status information:\r\nSubscriptions, etc...");
-                    break;
+                //saving command
+                switch (command[0])
+                {
+                    case "Unfreeze":
+                        Console.Out.WriteLine("Unfreezing");
+                        ProcessFrozenListCommands();
+                        Status = Status.Unfrozen;
+                        FrozenStateList.Clear();
+                        break;
+                    case "Freeze":
 
-                case "Crash":
-                    Console.WriteLine("Crashing");
-                    Process.GetCurrentProcess().Kill();
-                    break;
+                        Console.ForegroundColor = ConsoleColor.Cyan;
+                        Console.BackgroundColor = ConsoleColor.Black;
+                        Console.Out.WriteLine("I'm already frozen!");
+                        Console.ResetColor();
+                        break;
+                    default:
+                        FrozenStateList.Add(command);
+                        break;
+                }
 
-                case "Freeze":
-                    Console.Out.WriteLine("Freezing");
-                    Status = Status.Frozen;
-                    break;
+            }
+            else
+            {
+                switch (command[0])
+                {
+                    case "Status":
+                        Console.Out.WriteLine("Printing status information:\r\nSubscriptions, etc...");
+                        break;
 
-                case "Unfreeze":
-                    Console.Out.WriteLine("Unfreezing");
-                    Status = Status.Unfrozen;
-                    break;
+                    case "Crash":
+                        Console.WriteLine("Crashing");
+                        Process.GetCurrentProcess().Kill();
+                        break;
+
+                    case "Freeze":
+                        Console.ForegroundColor = ConsoleColor.Cyan;
+                        Console.BackgroundColor = ConsoleColor.Black;
+                        Console.Out.WriteLine("Freezing");
+                        Console.ResetColor();
+                        Status = Status.Frozen;
+                        break;
+
+                    case "Unfreeze":
+                        Console.Out.WriteLine("Unfreezing");
+                        Status = Status.Unfrozen;
+                        break;
+
+                    default:
+                        Console.Out.WriteLine("Command: " + command[0] + " doesn't exist!");
+                        break;
+                }
             }
         }
 
@@ -89,4 +146,5 @@ namespace CommonTypes
             return null;
         }
     }
+
 }
